Add game logic and fix OOP
Introduced the game logic, it is still missing out on some basic
features such as inform the end of a round as well as have basic
understanding of a player's hand and/or dealer's hand so as not
to ask for their choice of ace every time they get one in their
hand. Could possibly add inheritance to dealer and player class.
Overall it seems it has got a solid foundation and lacks on aesthetic
features to enance the user expereince.