Fix hit/stay option for player
Logic was kinda messy but now it is hopefully fixed and next step is
to make it look nice which is not necessarily needed but it is a nice
addition. Gotta add function quotes called docstrings for practice
cause that does not sound interesting at all. Sorry for these last
commit's body messages - just practicing stuff.